refactor: :truck: rinomina pagine
rinomino pagine con Page finale
